requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.3.9
cffi==1.12.3
cryptography==2.7
emoji==0.5.2
ephem==3.7.6.0
future==0.17.1
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
PySocks==1.7.0
python-telegram-bot==11.1.0
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.1
